Addressing Comments Structure changes
