+++ libraries/c_sdk/standard/mqtt/CMakeLists.txt
@@ -36,6 +36,7 @@ afr_module_sources(
         "${src_dir}/iot_mqtt_validate.c"
         "${src_dir}/iot_mqtt_context_connection.c"
         "${src_dir}/iot_mqtt_serializer_deserializer_wrapper.c"
+        "${src_dir}/iot_mqtt_managed_api.c"
         ${extra_mqtt_sources}
 )
